Saidas ajustando view
